Add floating point volume handling to AudioMixer
Use floating point volume in AudioMixer mixing when floating
point input is used with the new mixer engine.
AudioResampler is updated to take floating point volume to match.
Both legacy integer and floating point mixer engines work.

For now, integer volume is used when the new mixer engine
runs in integer input mode, for backward compatibility with
the legacy mixer.  The new mixer engine will generally run in
floating point input mode. When the legacy path is removed,
the integer volumes will be removed.

+++ services/audioflinger/AudioResamplerDyn.h
@@ -50,7 +50,7 @@ public:
 
     virtual void setSampleRate(int32_t inSampleRate);
 
-    virtual void setVolume(int16_t left, int16_t right);
+    virtual void setVolume(float left, float right);
 
     virtual void resample(int32_t* out, size_t outFrameCount,
             AudioBufferProvider* provider);
